+++ app/definitions/IThread.ts
@@ -1,36 +1,9 @@
 import Model from '@nozbe/watermelondb/Model';
-import { MarkdownAST } from '@rocket.chat/message-parser';
 
 import { IAttachment } from './IAttachment';
 import { IEditedBy, IUserChannel, IUserMention, IUserMessage, MessageType } from './IMessage';
 import { IReaction } from './IReaction';
 import { IUrl } from './IUrl';
-
-interface IFileThread {
-	_id: string;
-	name: string;
-	type: string;
-}
-
-export interface IThreadResult {
-	_id: string;
-	rid: string;
-	ts: string | Date;
-	msg?: string;
-	file?: IFileThread;
-	files?: IFileThread[];
-	groupable?: boolean;
-	attachments?: IAttachment[];
-	md?: MarkdownAST;
-	u: IUserMessage;
-	_updatedAt: Date;
-	urls?: IUrl[];
-	mentions?: IUserMention[];
-	channels?: IUserChannel[];
-	replies?: string[];
-	tcount?: number;
-	tlm?: Date;
-}
 
 export interface IThread {
 	id: string;
