+++ app/views/ThreadMessagesView/index.tsx
@@ -7,7 +7,6 @@ import { EdgeInsets, withSafeAreaInsets } from 'react-native-safe-area-context';
 import { HeaderBackButton, StackNavigationOptions, StackNavigationProp } from '@react-navigation/stack';
 import { RouteProp } from '@react-navigation/native';
 import { Observable, Subscription } from 'rxjs';
-import Database from '@nozbe/watermelondb/Database';
 
 import ActivityIndicator from '../../containers/ActivityIndicator';
 import I18n from '../../i18n';
@@ -33,13 +32,12 @@ import EventEmitter from '../../utils/events';
 import { LISTENER } from '../../containers/Toast';
 import SearchHeader from '../../containers/SearchHeader';
 import { ChatsStackParamList } from '../../stacks/types';
-import { IThreadResult, TThreadModel } from '../../definitions/IThread';
 import { Filter } from './filters';
 import DropdownItemHeader from './Dropdown/DropdownItemHeader';
 import Dropdown from './Dropdown';
 import Item from './Item';
 import styles from './styles';
-import { SubscriptionType, TSubscriptionModel } from '../../definitions/ISubscription';
+import { IMessage, SubscriptionType, TSubscriptionModel, TThreadModel } from '../../definitions';
 
 const API_FETCH_COUNT = 50;
 
@@ -259,8 +257,8 @@ class ThreadMessagesView extends React.Component<IThreadMessagesViewProps, IThre
 		remove,
 		lastThreadSync
 	}: {
-		update: IThreadResult[];
-		remove?: IThreadResult[];
+		update: IMessage[];
+		remove?: IMessage[];
 		lastThreadSync: Date;
 	}) => {
 		const { subscription } = this.state;
@@ -272,11 +270,9 @@ class ThreadMessagesView extends React.Component<IThreadMessagesViewProps, IThre
 		}
 
 		try {
-			const db: Database = database.active;
+			const db = database.active;
 			const threadsCollection = db.get('threads');
-			// TODO: Refactor when migrate room
-			// @ts-ignore
-			const allThreadsRecords = (await subscription.threads.fetch()) as TThreadModel[];
+			const allThreadsRecords = await subscription.threads.fetch();
 			let threadsToCreate: any[] = [];
 			let threadsToUpdate: any[] = [];
 			let threadsToDelete: any[] = [];
